fix: project arquitecture
