+++ src/ClothingApp.jsx
@@ -1,11 +1,10 @@
-import { UserProvider } from "./auth/context/UserProvider"
-import { AppRouter } from "./router/AppRouter"
-
+import { UserProvider } from "./pages/auth/context/UserProvider";
+import { AppRouter } from "./router/AppRouter";
 
 export const ClothingApp = () => {
   return (
     <UserProvider>
-      <AppRouter/>
+      <AppRouter />
     </UserProvider>
-  )
-}
+  );
+};
