--- src/auth/pages/index.js
@@ -1,1 +1,0 @@
-export * from './LoginPage'
