+++ src/pages/auth/index.js
@@ -1,0 +1,2 @@
+export * from "./login";
+export * from "./register";
