+++ src/pages/category/KidsPage.jsx
@@ -1,5 +1,3 @@
-import { Stepper } from "../../clothing/components/Stepper";
-import CarouselTeenGirls from "../../clothing/components/CarouselTeenGirls";
 import { Card } from "primereact/card";
 
 export const KidsPage = () => {
@@ -7,10 +5,7 @@ export const KidsPage = () => {
     <div className="animate__animated animate__fadeIn">
       <h1 className="text-center">Teen Girl</h1>
 
-      <Card className="py-4 px-8">
-        <Stepper />
-        <CarouselTeenGirls />
-      </Card>
+      <Card className="py-4 px-8"></Card>
     </div>
   );
 };
