+++ src/router/ClothingRoutes.jsx
@@ -1,9 +1,9 @@
 import { Navigate, Route, Routes } from "react-router-dom";
-import { AdminRoutes } from "../../admin/routes/AdminRoutes";
-import { Layout } from "../../components/Layout";
+import { AdminRoutes } from "../admin/routes/AdminRoutes";
+import { Layout } from "../components/Layout";
 
-import { HomePage, CouponsPage, QuizzPage, Cart } from "../../pages";
-import { KidsPage, MenPage, WomenPage } from "../../pages/category";
+import { HomePage, CouponsPage, QuizzPage, Cart } from "../pages";
+import { KidsPage, MenPage, WomenPage } from "../pages/category";
 
 export const ClothingRoutes = () => {
   return (
