+++ src/services/auth/authenticateUser.js
